RagModule/requirements.txt
# RagModule Dependencies
# Core FastAPI and web framework
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-multipart==0.0.12

# Database and MongoDB
pymongo==4.6.0
motor==3.3.2

# AI and ML
google-generativeai==0.3.2
numpy==1.26.4
scipy==1.11.4
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
chromadb==0.5.0

# Environment and configuration
python-dotenv==1.0.1

# Logging and utilities
loguru==0.7.2

# HTTP requests
requests==2.32.3

# Data processing
pandas==2.1.4

# Type hints
typing-extensions==4.8.0

# CORS middleware
starlette==0.27.0